refactor(south-korea): split parser out of south_korea_station_service (Refs #563)
- south_korea_station_service.dart: 394 -> 215 LOC (service shell only).
- lib/core/services/impl/south_korea_response_parser.dart (new): OPINET product-code map, station accumulator, brand/double parsing.
- SouthKoreaStationService keeps fuelForProductCode delegator + parseSingleProductResponse instance method so tests stayed unchanged.
